Port refactor from KJS
