+++ www/js/lib/uiUtil.js
@@ -1421,439 +1421,6 @@ function lockDisplayOrientation (val) {
 }
 
 /**
- * Parses a linked article in a loaded document in order to extract the first main paragraph (the 'lede') and first
- * main image (if any). This function currently only parses Wikimedia articles. It returns an HTML string, formatted
- * for display in a popover
- * @param {String} href The href of the article link from which to extract the lede
- * @param {String} baseUrl The base URL of the currently loaded article
- * @param {Document} articleDocument The DOM of the currently loaded article
- * @param {ZIMArchive} archive The archive from which to extract the lede
- * @returns {Promise<String>} A Promise for the linked article's lede HTML including first main image URL if any
- */
-function getArticleLede (href, baseUrl, articleDocument, archive) {
-    const uriComponent = removeUrlParameters(href);
-    const zimURL = deriveZimUrlFromRelativeUrl(uriComponent, baseUrl);
-    console.debug('Previewing ' + zimURL);
-    // Do a binary search in the URL index to get the directory entry for the requested article
-    return archive.getDirEntryByPath(zimURL).then(function (dirEntry) {
-        const readArticle = function (dirEntry) {
-            // Wrap legacy callback-based code in a Promise
-            return new Promise((resolve, reject) => {
-                // As we're reading Wikipedia articles, we can assume that they are UTF-8 encoded HTML data
-                archive.readUtf8File(dirEntry, function (fileDirEntry, htmlArticle) {
-                    const parser = new DOMParser();
-                    const doc = parser.parseFromString(htmlArticle, 'text/html');
-                    const articleBody = doc.body;
-                    if (articleBody) {
-                        // Establish the popup balloon's base URL and the absolute path for calculating the ZIM URL of links and images
-                        const balloonBaseURL = encodeURI(fileDirEntry.namespace + '/' + fileDirEntry.url.replace(/[^/]+$/, ''));
-                        const docUrl = new URL(articleDocument.location.href);
-                        const rootRelativePathPrefix = docUrl.pathname.replace(/([^.]\.zim\w?\w?\/).+$/i, '$1');
-                        let balloonString = '';
-                        // Remove all standalone style elements, because their content is shown by both innerText and textContent
-                        const styleElements = Array.from(articleBody.querySelectorAll('style'));
-                        styleElements.forEach(style => {
-                            style.parentNode.removeChild(style);
-                        });
-                        const paragraphs = Array.from(articleBody.querySelectorAll('p'));
-                        // Filter out empty paragraphs or those with less than 50 characters
-                        const nonEmptyParagraphs = paragraphs.filter(para => {
-                            const text = para.innerText.trim();
-                            return !/^\s*$/.test(text) && text.length >= 50;
-                        });
-                        if (nonEmptyParagraphs.length > 0) {
-                            let cumulativeCharCount = 0;
-                            // Add enough paras to complete the word count
-                            for (let i = 0; i < nonEmptyParagraphs.length; i++) {
-                                // Get the character count: to fill the larger box we need ~850 characters (815 plus leeway)
-                                const plainText = nonEmptyParagraphs[i].innerText;
-                                cumulativeCharCount += plainText.length;
-                                // In ServiceWorker mode, we need to transform the URLs of any links in the paragraph
-                                if (params.contentInjectionMode === 'serviceworker') {
-                                    const links = Array.from(nonEmptyParagraphs[i].querySelectorAll('a'));
-                                    links.forEach(link => {
-                                        const href = link.getAttribute('href');
-                                        if (href && !/^#/.test(href)) {
-                                            const zimURL = deriveZimUrlFromRelativeUrl(href, balloonBaseURL);
-                                            link.href = rootRelativePathPrefix + encodeURI(zimURL);
-                                        }
-                                    });
-                                }
-                                // Get the transformed HTML. Note that in Safe mode, we risk breaking the UI if user clicks on an
-                                // embedded link, so only use innerText in that case
-                                const content = params.contentInjectionMode === 'jquery' ? plainText
-                                    : nonEmptyParagraphs[i].innerHTML;
-                                balloonString += '<p>' + content + '</p>';
-                                // If we have enough characters to fill the box, break
-                                if (cumulativeCharCount >= 850) break;
-                            }
-                        }
-                        // If we have a lede, we can now add an image to the balloon
-                        const images = articleBody.querySelectorAll('img');
-                        let firstImage = null;
-                        if (images && params.contentInjectionMode === 'serviceworker') {
-                            // Iterate over images until we find one with a width greater than 50 pixels
-                            // (this filters out small icons)
-                            const imageArray = Array.from(images);
-                            for (let j = 0; j < imageArray.length; j++) {
-                                if (imageArray[j] && imageArray[j].width > 50) {
-                                    firstImage = imageArray[j];
-                                    break;
-                                }
-                            }
-                        }
-                        if (firstImage) {
-                            // Calculate root relative URL of image
-                            const imageZimURL = encodeURI(deriveZimUrlFromRelativeUrl(firstImage.getAttribute('src'), balloonBaseURL));
-                            firstImage.src = rootRelativePathPrefix + imageZimURL;
-                            balloonString = firstImage.outerHTML + balloonString;
-                        }
-                        if (!balloonString) {
-                            reject(new Error('No article lede or image'));
-                        } else {
-                            resolve(balloonString);
-                        }
-                    } else {
-                        reject(new Error('No article body found'));
-                    }
-                });
-            });
-        }
-        if (!dirEntry) {
-            return Promise.reject(new Error('No directory entry found'));
-        } else if (dirEntry.redirect) {
-            // If the dirEntry is a redirect, we need to resolve it before reading the article
-            return new Promise((resolve, reject) => {
-                archive.resolveRedirect(dirEntry, function (reDirEntry) {
-                    resolve(readArticle(reDirEntry));
-                });
-            }).catch(error => {
-                return Promise.reject(error);
-            });
-        } else {
-            // Directory entry was found, so now read the article data
-            return Promise.resolve(readArticle(dirEntry));
-        }
-    }).catch(function (err) {
-        throw new Error('Could not get Directory Entry for ' + zimURL, err);
-    });
-}
-
-/**
- * A function to attach the tooltip CSS for popovers (NB this does not attach the box itself, only the CSS)
- * @param {Document} doc The document to which to attach the popover stylesheet
- * @param {Boolean} dark An optional parameter to adjust the background colour for dark themes (generally not needed for inversion-based themes)
- */
-function attachKiwixPopoverCss (doc, dark) {
-    const colour = dark ? '#darkgray' : '#black';
-    const backgroundColour = dark ? '#111' : '#ebf4fb';
-    insertLinkElement(doc, `
-        .kiwixtooltip {
-            position: absolute;
-            bottom: 1em;
-            /* prettify */
-            padding: 0 5px 5px;
-            color: ${colour};
-            background: ${backgroundColour};
-            border: 0.1em solid #b7ddf2;
-            /* round the corners */
-            border-radius: 0.5em;
-            /* handle overflow */
-            overflow: visible;
-            text-overflow: ellipsis;
-            /* handle text wrap */
-            overflow-wrap: break-word;
-            word-wrap: break-word;
-            /* add fade-in transition */
-            opacity: 0;
-            transition: opacity 0.3s;
-        }
-        
-        .kiwixtooltip img {
-            float: right;
-            margin-left: 5px;
-            max-width: 40%;
-            height: auto;
-        }
-
-        #popcloseicon {
-            padding-top: 1px;
-            padding-right: 2px;
-            font-size: 20px;
-            font-family: sans-serif;
-        }
-        
-        #popcloseicon:hover { 
-            cursor: pointer;
-        }
-        
-        #popbreakouticon {
-            height: 18px;
-            margin-right: 18px;
-        }
-        
-        #popbreakouticon:hover {
-            cursor: pointer;
-        }
-        
-        /* Prevent native iOS popover on Safari if option is enabled */
-        body { 
-            -webkit-touch-callout: none !important;
-        }
-        `,
-        // The id of the style element for easy manipulation
-        'kiwixtooltipstylesheet'
-    );
-}
-
-/**
- * Attaches a popover div for the given link to the given document's DOM
- * @param {Event} ev The event which has fired this popover action
- * @param {Element} link The link element that is being actioned
- * @param {String} articleBaseUrl The base URL of the currently loaded document
- * @param {Boolean} dark An optional value to switch colour theme to dark if true
- * @param {ZIMArchive} archive The archive from which the popover information is extracted
- * @returns {Promise<div>} A Promise for the attached popover div or undefined if the popover is not attached
- */
-function populateKiwixPopoverDiv (ev, link, articleBaseUrl, dark, archive) {
-    // Do not show popover if the user has initiated an article load (set in filterClickEvent)
-    if (link.articleisloading || link.popoverisloading) return Promise.resolve();
-    const linkHref = link.getAttribute('href');
-    // Do not show popover if there is no href or with certain landing pages
-    if (!linkHref || /^wikivoyage/i.test(archive.file.name) &&
-      (appstate.expectedArticleURLToBeDisplayed === archive.landingPageUrl ||
-      appstate.expectedArticleURLToBeDisplayed === 'A/Wikivoyage:Offline_reader_Expedition/Home_page')) {
-        return Promise.resolve();
-    }
-    link.popoverisloading = true;
-    // Do not display a popover if one is already showing for the current link
-    const kiwixPopover = ev.target.ownerDocument.querySelector('.kiwixtooltip');
-    if (kiwixPopover && kiwixPopover.dataset.href === linkHref) return Promise.resolve();
-    // console.debug('Attaching popover...');
-    const currentDocument = ev.target.ownerDocument;
-    const articleWindow = currentDocument.defaultView;
-    // Remove any existing popover(s) that the user may not have closed before creating a new one
-    removeKiwixPopoverDivs(currentDocument);
-    setTimeout(function () {
-        // Check if the user has moved away from the link or has clicked it, and abort display of popover if so
-        if (link.articleisloading || !link.matches(':hover') && !link.touched && currentDocument.activeElement !== link) {
-            // console.debug('Aborting popover display for ' + linkHref + ' because user has moved away from link or clicked it');
-            link.popoverisloading = false;
-            return Promise.resolve();
-        }
-        // Create a new Kiwix popover container
-        const divWithArrow = createNewKiwixPopoverCointainer(articleWindow, link, ev);
-        const div = divWithArrow.div;
-        const span = divWithArrow.span;
-        // Get the article's 'lede' (first main paragraph or two) and the first main image (if any)
-        return getArticleLede(linkHref, articleBaseUrl, currentDocument, archive).then(function (html) {
-            div.style.justifyContent = '';
-            div.style.alignItems = '';
-            div.style.display = 'block';
-            const breakoutIconFile = window.location.pathname.replace(/\/[^/]*$/, '') + (dark ? '/img/icons/new_window_white.svg' : '/img/icons/new_window_black.svg');
-            const backgroundColour = '#ebf4fb';
-            // DEV: Most style declarations in this div only work properly inline. If added in stylesheet, even with !important, the positioning goes awry
-            // (appears to be a timing issue related to the reservation of space given that the div is inserted dynamically).
-            div.innerHTML = `<div style="position: relative; overflow: hidden; height: ${div.style.height};">
-                <div style="background: ${backgroundColour} !important; opacity: 70%; position: absolute; top: 0; right: 0; display: flex; align-items: center; padding: 0; z-index: 1;">
-                    <img id="popbreakouticon" src="${breakoutIconFile}" />
-                    <span id="popcloseicon">X</span>
-                </div>
-                <div style="padding-top: 3px">${html}</div>
-            </div>`;
-            // Now it is populated, we can attach the arrow to the div
-            div.appendChild(span);
-            // Programme the icons
-            addEventListenersToPopoverIcons(link, div, currentDocument);
-            setTimeout(function () {
-                div.popoverisloading = false;
-            }, 900);
-            // Finally, return the div DOM element
-            return div;
-        }).catch(function (err) {
-            console.warn(err);
-            // Remove the div
-            div.style.opacity = '0';
-            div.parentElement.removeChild(div);
-            link.dataset.touchevoked = false;
-            link.popoverisloading = false;
-        });
-    }, 600);
-}
-
-/**
- * Create a new empty Kiwix popover container and attach it to the current document appropriately sized and positioned
- * in relation to the given anchor and available screen width and height. Also returns the arrow span element which can be
- * attached to the div after the div is populated with content.
- * @param {Window} win The window of the article DOM
- * @param {Element} anchor The anchor element that is being actioned
- * @param {Event} event The event which has fired this popover action
- * @returns {Object} An object containing the popover div and the arrow span elements
- */
-function createNewKiwixPopoverCointainer (win, anchor, event) {
-    const div = document.createElement('div');
-    const linkHref = anchor.getAttribute('href');
-    const currentDocument = win.document;
-    div.popoverisloading = true;
-    const screenWidth = win.innerWidth - 40;
-    const screenHeight = document.documentElement.clientHeight;
-    let margin = 40;
-    let divWidth = 512;
-    if (screenWidth <= divWidth) {
-        divWidth = screenWidth;
-        margin = 10;
-    }
-    // Check if we have restricted screen height
-    const divHeight = screenHeight < 512 ? 160 : 256;
-    div.style.width = divWidth + 'px';
-    div.style.height = divHeight + 'px';
-    div.style.display = 'flex';
-    div.style.justifyContent = 'center';
-    div.style.alignItems = 'center';
-    div.className = 'kiwixtooltip';
-    div.innerHTML = '<p>Loading ...</p>';
-    div.dataset.href = linkHref;
-    // DEV: We need to insert the div into the target document before we can obtain its computed dimensions accurately
-    currentDocument.body.appendChild(div);
-    // Calculate the position of the link that is being hovered
-    const linkRect = anchor.getBoundingClientRect();
-    // Initially position the div 20px above the link
-    let triangleDirection = 'top';
-    const divOffsetHeight = /UWP/.test(params.appType) ? div.offsetHeight * params.relativeFontSize / 100 + 20 : div.offsetHeight + 20;
-    let divRectY = linkRect.top - divOffsetHeight;
-    if (/UWP/.test(params.appType)) divRectY = divRectY * 100 / params.relativeFontSize;
-    let triangleY = divHeight + 6;
-    // If we're less than half margin from the top, move the div below the link
-    if (divRectY < margin / 2) {
-        triangleDirection = 'bottom';
-        divRectY = linkRect.bottom + 20;
-        triangleY = -16;
-        if (/UWP/.test(params.appType)) divRectY = divRectY * 100 / params.relativeFontSize;
-    }
-    // Position it horizontally in relation to the pointer position
-    let divRectX, triangleX;
-    if (event.type === 'touchstart') {
-        divRectX = event.touches[0].clientX - divWidth / 2;
-        triangleX = event.touches[0].clientX - divRectX - 20;
-    } else if (event.type === 'focus') {
-            divRectX = linkRect.left + linkRect.width / 2 - divWidth / 2;
-            triangleX = linkRect.left + linkRect.width / 2 - divRectX - 20;
-    } else {
-        divRectX = event.clientX - divWidth / 2;
-        triangleX = event.clientX - divRectX - 20;
-    }
-    // If right edge of div is greater than margin from the right side of window, shift it to margin
-    if (divRectX + divWidth * params.relativeFontSize / 100 > screenWidth - margin) {
-        triangleX += divRectX;
-        divRectX = screenWidth - divWidth * params.relativeFontSize / 100 - margin;
-        triangleX -= divRectX;
-    }
-    // If we're less than margin to the left, shift it to margin px from left
-    if (divRectX * params.relativeFontSize / 100 < margin) {
-        triangleX += divRectX;
-        divRectX = margin;
-        triangleX -= divRectX;
-    }
-    // Adjust triangleX if necessary
-    if (triangleX < 10) triangleX = 10;
-    if (triangleX > divWidth - 10) triangleX = divWidth - 10;
-    // Adjust positions to take into account the font zoom factor
-    divRectX = divRectX * 100 / params.relativeFontSize;
-    triangleX = triangleX * 100 / params.relativeFontSize;
-    var adjustedScrollY = win.scrollY * 100 / params.relativeFontSize;
-    // Now set the calculated x and y positions, taking into account the zoom factor
-    div.style.top = divRectY + adjustedScrollY + 'px';
-    div.style.left = divRectX + 'px';
-    div.style.opacity = '1';
-    // Now create the arrow span element. Note that we cannot attach it yet as we need to populate the div first
-    // and doing so will overwrite the innerHTML of the div
-    const triangleColour = '#b7ddf2'; // Same as border colour of div
-    const span = document.createElement('span');
-    span.style.cssText = `
-        width: 0;
-        height: 0;
-        border-${triangleDirection}: 16px solid ${triangleColour};
-        border-left: 8px solid transparent !important;
-        border-right: 8px solid transparent !important;
-        position: absolute;
-        top: ${triangleY}px;
-        left: ${triangleX}px;
-    `;
-    return { div: div, span: span };
-}
-
-/**
- * Adds event listeners to the popover's control icons
- * @param {Element} anchor The anchor which launched the popover
- * @param {Element} popover The containing element of the popover (div)
- * @param {Document} doc The doucment on which to operate
- */
-function addEventListenersToPopoverIcons (anchor, popover, doc) {
-    const breakout = function (e) {
-        e.preventDefault();
-        e.stopPropagation();
-        // Adding the newcontainer property to the anchor will be cauught by the filterClickEvent function and will open in new tab
-        anchor.newcontainer = true;
-        anchor.click();
-        closePopover(popover);
-    }
-    const closeIcon = doc.getElementById('popcloseicon');
-    const breakoutIcon = doc.getElementById('popbreakouticon');
-    // Register mousedown event (should work in all contexts)
-    closeIcon.addEventListener('mousedown', function () {
-        closePopover(popover);
-    }, true);
-    breakoutIcon.addEventListener('mousedown', breakout, true);
-    // Additionally register either pointerdown or touchstart if supported for faster response
-    const eventName = window.PointerEvent ? 'pointerdown' : 'touchstart';
-    closeIcon.addEventListener(eventName, function (e) {
-        e.preventDefault();
-        e.stopPropagation();
-        closePopover(popover);
-    }, true);
-    breakoutIcon.addEventListener(eventName, breakout, true);
-}
-
-/**
- * Remove any preview popover DIVs found in the given document
- * @param {Document} doc The document from which to remove any popovers
- */
-function removeKiwixPopoverDivs (doc) {
-    const divs = doc.getElementsByClassName('kiwixtooltip');
-    setTimeout(function () {
-        // Gather any popover divs (on rare occasions, more than one may be displayed)
-        Array.prototype.slice.call(divs).forEach(function (div) {
-            // Do not remove any popover in process of loading
-            if (div.popoverisloading) return;
-            let timeoutID;
-            const fadeOutDiv = function () {
-                clearTimeout(timeoutID);
-                // Do not close any div which is being hovered
-                if (!div.matches(':hover')) {
-                    closePopover(div);
-                } else {
-                    timeoutID = setTimeout(fadeOutDiv, 250);
-                }
-            };
-            timeoutID = setTimeout(fadeOutDiv, 0);
-        });
-    }, 400);
-}
-
-/**
- * Closes the specified popover div, with fadeout effect, and removes it from the DOM
- * @param {Element} div The div to close
- */
-function closePopover (div) {
-    div.style.opacity = '0';
-    setTimeout(function () {
-        if (div && div.parentElement) {
-            div.parentElement.removeChild(div);
-        }
-    }, 200);
-};
-
-/**
  * Finds the closest <a> or <area> enclosing tag of an element.
  * Returns undefined if there isn't any.
  *
@@ -1914,9 +1481,6 @@ export default {
     initTouchZoom: initTouchZoom,
     appIsFullScreen: appIsFullScreen,
     lockDisplayOrientation: lockDisplayOrientation,
-    attachKiwixPopoverCss: attachKiwixPopoverCss,
-    populateKiwixPopoverDiv: populateKiwixPopoverDiv,
-    removeKiwixPopoverDivs: removeKiwixPopoverDivs,
     reportAssemblerErrorToAPIStatusPanel: reportAssemblerErrorToAPIStatusPanel,
     reportSearchProviderToAPIStatusPanel: reportSearchProviderToAPIStatusPanel,
     warnAndOpenExternalLinkInNewTab: warnAndOpenExternalLinkInNewTab,
